[Upd] weather is supposedly good
